Add evas_object_table, make evas_object_box more consistent.
Table code is still *incomplete*, it just do homogeneous layouts as
I'm still trying to figure out how to make it great.

I'm not expecting to make layout configurable, as we did for box, but
if you think it's required we can do that later.

Now that the public API of both BOX and TABLE are in, we can add these
as parts of Edje.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk/evas@37359 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33