Pending patch from glima, who's on vacations.
Add two new canvas level callbacks: OBJECT_FOCUS_IN/OUT
As we already had callbacks for objects gaining or losing focus, then
two more for the Canvas, now we can have the entire Evas be notified when
any object changes its focused status. The object in question is passed
as the event_info for the callback.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk/evas@52196 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33