Evas font-engine: Fixed kerning issue and a major rendering bug.
Kerning: We are walking the string visually so we don't need to do
anything special for kerning when in rtl, freetype works with "left" and
"right" which we automatically get.

Rendering bug: Cedric found that in some cases there were missing
characters. This was caused because I forgot to convert the kerning from
16.6 fixed point to int.

git-svn-id: http://svn.enlightenment.org/svn/e/trunk/evas@58783 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33