From: Jiyoun Park <jy0703.park@samsung.com> Subject: [E-devel] [Patch] modify gl engine's animated function related with cache entry
I modified the gl engine code related with animated images
This is very trivial.  Evas image object passes images to the engine. 
In the software engine, it is a cache entry , but in the GL engine, it is 
an Evas_GL Image. So I modified the gl engine code to get the cache entry
from the gl image.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk/evas@64143 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33