From: "Sung W. Park" <sungwoo@gmail.com> Subject: [E-devel] [E-Devel][Patch] Evas GL Color Format Enum change (and ElmGLview changes accordingly)
I'm submitting a patch that changes the color format for Evas GL.

When I first wrote Evas_GL, I just had EVAS_GL_RGB_8 and EVAS_GL_RGBA_8 and etc
but it was misleading for some people.  It was more of a filler since I couldn't decide on
a name.  I'm finally changing it to make it more clear.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk/evas@64491 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33