From: Jérôme Pinot <ngc891@gmail.com> Subject: [E-devel] [patch] evas doxygen doc (1)
There are a lot of small things to fix in the evas doc. Here is a first
batch. Fixes are trivial and obvious. Just a question about the first
hunk, I removed the <tdb> that replaced valid email for companies
because it makes doxygen complain about the unknown tag and, well, it
doesn't give any information either. Was there an explanation about
adding this ?

Joined is the patch along with the doc build log diff.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk/evas@69749 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33