Evas: Fix bugs in Wayland Engine(s) with regard to object geometry & position. Fix bug when clipping to viewport/framespace. These changes fix Several buggers on Trac which related to the Wayland Engine(s).
Fix minor bug in evas_render where clipping for framespace would
not take into account the frame height and also move the frame clip to
the values specified in framespace (not viewport).

Fix buggers in evas_object_main which was causing objects (when created using 
the wayland engines) to be an incorrect size & position (objects will now
be the same size/position in EFL Wayland as they are in X11).

Fix evas_object_geometry_get to return values based on where they were
moved (in relation to framespace). (These fixes are for the Wayland
Engines and do not affect X11).

Remove nw/nh from evas_object_resize (not needed variables anymore).



git-svn-id: http://svn.enlightenment.org/svn/e/trunk/evas@74565 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33